+++ include/lh/interval/bounds/initializer.h
@@ -1,0 +1,9 @@
+#ifndef LH_INTERVAL_BOUNDS_INITIALIZER_H
+#define LH_INTERVAL_BOUNDS_INITIALIZER_H
+
+#include "lh/initializer.h"
+
+#define lh_interval_bounds_initializer(lower_value, upper_value)               \
+    lh_initializer(lower_value, upper_value)
+
+#endif /* LH_INTERVAL_BOUNDS_INITIALIZER_H */